refactor: structure project
