+++ src/main/java/com/artocons/carshop/config/CarShopConfiguration.java
@@ -1,10 +1,7 @@
 package com.artocons.carshop.config;
 
 import com.artocons.carshop.util.CarShopConstants;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
@@ -19,7 +16,7 @@ public class CarShopConfiguration implements WebMvcConfigurer {
 
     @Override
     public void addViewControllers(ViewControllerRegistry registry) {
-        registry.addViewController(CarShopConstants.LOGIN_PATH).setViewName(LOGIN_PAGE);
+        registry.addViewController(CarShopConstants.AUTH_PATH).setViewName(LOGIN_PAGE);
         registry.addViewController(CarShopConstants.ADMIN_PATH).setViewName(ADMIN_PANEL_PAGE);
         registry.addViewController(CarShopConstants.CART_PATH).setViewName(CART_PAGE);
         registry.addViewController(CarShopConstants.ORDER_PATH).setViewName(ORDER_PAGE);
