+++ src/main/java/com/artocons/carshop/controller/page/CarListPageController.java
@@ -3,10 +3,10 @@ package com.artocons.carshop.controller.page;
 import com.artocons.carshop.exception.ResourceNotFoundException;
 import com.artocons.carshop.exception.ResourceVaidationException;
 import com.artocons.carshop.persistence.model.*;
+import com.artocons.carshop.persistence.request.AjaxRequest;
+import com.artocons.carshop.persistence.response.AjaxResponse;
 import com.artocons.carshop.service.CarService;
 import com.artocons.carshop.service.CartService;
-import com.artocons.carshop.service.StockService;
-import com.artocons.carshop.validation.QuantityValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.*;
 import org.springframework.http.HttpStatus;
