+++ src/main/java/com/artocons/carshop/util/CarShopConstants.java
@@ -5,7 +5,7 @@ public class CarShopConstants {
     public static final String ROOT_PATH = "/";
     public static final String CARS_PATH = "/cars";
     public static final String ADMIN_PATH = "/admin";
-    public static final String LOGIN_PATH = "/login";
+    public static final String AUTH_PATH = "/auth";
     public static final String CART_PATH = "/cart";
     public static final String PRODUCT_PATH = "/product";
     public static final String ORDER_PATH = "/order";
