+++ src/main/webapp/WEB-INF/tags/common/login.tag
@@ -1,8 +1,8 @@
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
-<c:url var="loginLink" value="/login"/>
+<c:url var="authLink" value="/auth"/>
 
     <div class="btn-group" role="group" aria-label="Second group">
-        <a href="<c:url value="${loginLink}"/>">
+        <a href="<c:url value="${authLink}"/>">
             <button type="button" class="btn btn-primary btn-sm">Login</button>
         </a>
     </div>
